+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 <!-- do not remove -->
 
+## 0.0.4
+
+- Add CLI flags for filtering by type, repo, etc
+- Use ghapi.event parallel event stream
+
 ## 0.0.3
 
 - failed to handle stored token